requirements.txt
alabaster==0.7.12
amqp==2.4.1
argon2-cffi==19.1.0
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
decorator==4.3.2
defusedxml==0.5.0
Django==2.0.13
django-allauth==0.38.0
django-anymail==6.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.5
django-model-utils==3.1.2
django-redis==4.10.0
djangorestframework==3.9.1
djangorestframework-simplejwt==4.0.0
docutils==0.14
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.2
flake8==3.7.5
idna==2.8
imagesize==1.1.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.1
mccabe==0.6.1
model-mommy==1.6.0
more-itertools==6.0.0
mypy==0.670
mypy-extensions==0.4.1
oauthlib==3.0.1
packaging==19.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
prompt-toolkit==2.0.9
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.3.1
pytest==4.2.0
pytest-django==3.4.7
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.9
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
snowballstemmer==1.2.1
sparkpost==1.3.6
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.3.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==4.1.2
